+++ src/boole/index.ts
@@ -1,0 +1,13 @@
+export type { BooleFormula, FormulaClassification, TruthAssignment, TruthTableRow } from './ast'
+export { BooleError } from './ast'
+export {
+  analyzeEntailment,
+  analyzeEquivalence,
+  analyzeFormula,
+  collectVariables,
+  evaluateBooleFormula,
+  type EntailmentAnalysis,
+  type FormulaAnalysis,
+} from './evaluator'
+export { parseBooleFormula } from './parser'
+export { tokenizeBoole } from './tokenizer'